requirements.txt
beautifulsoup4==4.6.3
Django==2.1.3
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.2
django-cors-headers==2.4.0
django-libsass==0.7
django-livereload-server==0.3.2
django-progressive-web-app==0.1.1
djangorestframework==3.9.0
libsass==0.15.1
Pillow==5.3.0
pytz==2018.7
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
tornado==5.1.1
urllib3==1.24.1
